JWT: Enable reading a JSON Web Key Set from a file (#180)
* Add ForwardingSignatureValidator

ForwardingSignatureValidator holds a pointer to some other SignatureValidator
and will forward ValidateSignature requests to this validator.
The validator it references can be changed by calling Replace().

* Make JWKS a oneoff with either inline or file content

This changes the jwt.proto definition to take either an inline JWKS
struct, or a message containing a file path and a refresh interval.

The intention is that when a file path and refresh interval is provided,
we create a ForwardingSignatureValidator, and periodically update its internal
SignatureValidator with content from the referenced file.

When passing inline content, behavior should remain unchanged.

* Load JWKS from a file

When loading the JWT configuration, check if the config provides inline
JWKS content, or a reference to a file. If we get a reference to a file,
we set up a goroutine to periodically fetch the file and update a
ForwardingSignatureValidator.