Pushgateway: Add the ability to scrape targets
When running bb_worker on a system that isn't reachable externally, it
may be desirable to use the Pushgateway client to submit metrics
externally. In those cases it may also be useful to merge in metrics
coming from other processes such as Prometheus node exporter.

This change adds a configuration option for specifying a list of scrape
targets. Every time metrics are pushed to Pushgateway, it will scrape
these targets and extract the metrics matching a given regular
expression.