Refactor to resolve with path.Parser
Right now path.Resolve() takes UNIX style paths. In order to add Windows
pathanme support, one approach would be to have path.ResolveUNIX() and
path.ResolveWindows(). But this is not sufficient, because during a
single resolution pass we could encounter both kinds of paths.

We solve this by wrapping pathname strings in a path.Parser. This allows
path.Resolve() to remain a generic resolution algorithm that is
oblivious of the path that is being resolved.