proto/configuration: Add `RoutingBlobAccess`
This change adds a config option for `RoutingBlobAccess`, which is
intended to hold all possible ways to rewrite instance names.

While it is possible to affect instance name by (ab)using
`DemultiplexingBlobAccess`, there are some sharp edges (e.g. using it to
add an instance name may not work as expected, as not only must the
instance name be valid on the inner blobstore, but it must be valid once
returned upwards as well). `RoutingBlobAccess` will attempt to fix that
by offering easier-to-use semantics around instance name rewriting.