bazel: Add `.bazelversion` file
This change adds a `.bazelversion` file which allows `bazelisk` users to
use the expected version of `bazel` when building. The version of bazel
specified matches that of the one used in Github Actions currently.

This is a small quality-of-life improvement, as otherwise the default of
"latest release" will download bazel 7, fail to build, and create
bzlmod-related files in the process.