Add support for TLS to HTTP servers
We support TLS for gRPC clients, gRPC servers, and HTTP clients. HTTP
servers are the only ones for which TLS is not supported. This is
typically not an issue, because people generally put ingress controllers
in front of HTTP servers. However, if people want to run Buildbarn in
bare metal environments it may be useful to support this.