Only request TLS client certificates if needed
For gRPC we can always request that client certificates are sent, even
if we don't do actual certificate validation. For HTTP this is a bit
more problematic, because browsers tend to show popups/prompts whenever
the server asks for a certificate.

This change adds an argument to NewTLSConfigFromServerConfiguration() to
indicate that certificates should be sent over. We only enable this if
the server has a TLS client certificate validating policy in place.