Make termination of Buildbarn binaries a bit more reliable
If a Buildbarn binary is launched as part of a process group and is not
the process group leader, it may be the case that it receives signals
delivered to the process group which are ignored by the process itself.
This causes our final termination signal to never arrive.

Let's put a safety belt in place, calling os.Exit() if a sufficient
amount of time passes.