Add a utility function for converting path.Trace to a list
There are cases in which users of path.Trace want to do forward
iteration of the path. path.Trace does not permit it, because it's a
backward pointing linked list.

Instead of adding a ToList() function, we could have extended Trace to
implement Parser. But this is needlessly complex, because traces are
guaranteed to be well formed, and unlike Builder don't contain any ".."
components.