Depend on remote-apis as a Bazel module
remote-apis just switched to bzlmod. This change patches up bb-storage
to depend on it as a module, instead of pulling it in via Gazelle. Right
now this still requires us to apply a local patch, because the
go_proto_library() targets in remote-apis use the old gRPC v1 compiler
for generating sources. See this PR for more details:

https://github.com/bazelbuild/remote-apis/pull/308