Open directories using O_SEARCH
Right now we open directories using O_RDONLY. This is problematic,
because it means we can't perform operations on directories that only
have execute permissions. The solution for this is to use O_SEARCH,
which was added in POSIX 2008.

Linux doesn't offer O_SEARCH, but we can emulate it using O_PATH. macOS
has supported this feature since Ventura (October 2022).