Switch sharding struct from array to map
Improves the ability to perform resharding by switching the sharding
struct from array to map. Each map entry has a key which is used in
rendezvous hashing to deterministically select which shard to use from
the collection of keys. When a shard is removed it is guaranteed that
only blobs which belonged to the removed shard will resolve to a new
shard.

In combination with ReadFallbackConfigurations this allows adding and
removing shards with minimal need to rebalance the blobs between the
shards.

For more details and migration instructions, see:
https://github.com/buildbarn/bb-adrs/blob/bf2066633e1712a3ef7c295d37cd52e65867391d/0011-rendezvous-hashing.md