Remote authenticator and authorizer (#234)
* Move authorizer_factory out to pkg/auth/configuration

Prepare for remote authentication and authorization to avoid future
curcular dependency.

* Add grpcClientFactory to grpc.NewAuthenticatorFromConfiguration

* Add grpcClientFactory to http.NewAuthenticatorFromConfiguration

* Add grpcClientFactory to NewAuthorizerFromConfiguration

* Add remote auth .proto definitions

* Implement remote authenticator and authorizer

The authenticate and authorize tasks can now be sent remotely over gRPC
to an external service. This way, custom authentication and
authorization does not require a modified builds of the Buildbarn
components.

To avoid spamming the remote service with calls for every REv2 request
and keep the latency low, the verdicts, both allow and deny, are cached
for a duration specified in the response from the remote service.

* Deduplicate JWT header authenticator code