Improve windows mmap block performance (#264)
I noticed that the write performance was very poor on Windows when
running bb_worker. Specifically during the initial fetch of all of the
blobs from the CAS, bb_worker was spending a lot of time stuck in the
WriteAt function. Writing via the memory map reduced this back down to
almost nothing.