Add generic gRPC stream forwarding
To be able to pass Bazel's build event stream (BES) to the same DNS
name, without having to add an extra L7 router in front of the
bb-storage frontend, add a configuration to forward specific gRPC
methods to other backends. No authorization is possible on the passed
through messages because Buildbarn has no knowledge about the semantics
of the forwarded messages.

The gRPC reflection service has also been extended to forward requests
that cannot be resolved locally.